refactor: move type helpers into a separate file
